Crate libra_metrics

Modules

counters

metric_server

Structs

DurationHistogram
A small wrapper around Histogram to handle the special case of duration buckets. This Histogram will handle the correct granularty for logging time duration in a way that fits the used buckets.

Histogram
A Metric counts individual observations from an event or sample stream in configurable buckets. Similar to a summary, it also provides a sum of observations and an observation count.

OpMetrics

ServiceMetrics

Functions

dump_all_metrics_to_file_periodically

export_counter

get_all_metrics

get_metric_name

Type Definitions

IntCounter
The integer version of Counter. Provides better performance if metric values are all integers.

IntCounterVec
The integer version of CounterVec. Provides better performance if metric values are all integers.

IntGauge
The integer version of Gauge. Provides better performance if metric values are all integers.

IntGaugeVec
The integer version of GaugeVec. Provides better performance if metric values are all integers.